# dkim-lookup-app/public/openapi.yaml
openapi: 3.0.3
info:
  title: DKIM Registry API
  description: |-
    This is the API specification for the DKIM Registry API, which follows the OpenAPI 3.0.3 specification.
  version: "0.1"
servers:
  - url: https://registry.prove.email/api
  - url: /api
    description: Development
paths:
  /key:
    get:
      tags:
        - dkim-key
      summary: List DKIM keys
      description: Returns a list of DKIM keys for a given domain or domain suffix
      operationId: listKeys
      parameters:
        - name: domain
          in: query
          description: Match the domain name and any subdomains.
          required: true
          schema:
            type: string
          example: 'example.com'
      responses:
        '400':
          description: Missing parameter
        '429':
          description: Rate limit exceeded
        '500':
          description: Unexpected error
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DkimKey'
  /dsp:
    post:
      tags:
        - domain-selector-pair
      summary: Add a domain/selector pair
      description: Add a domain/selector pair
      operationId: addDsp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainSelectorPair'
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                required:
                  - added
                  - message
                properties:
                  added:
                    type: boolean
                  message:
                    $ref: '#/components/schemas/DomainSelectorPair'

        '400':
          description: Invalid input
        '500':
          description: Unexpected error
  /domains/{name}:
    get:
      tags:
        - dkim-key
      summary: List DKIM keys
      description: Returns a list of DKIM keys for a given domain or domain suffix
      operationId: listDomains
      deprecated: true
      parameters:
        - name: name
          in: path
          description: Match the domain name and any subdomains.
          required: true
          schema:
            type: string
          example: 'example.com'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DkimKey'
        '429':
          description: Rate limit exceeded
        '500':
          description: Unexpected error
components:
  schemas:
    DkimKey:
      type: object
      required:
        - value
        - domain
        - selector
      properties:
        value:
          type: string
          example: k=rsa; t=s; p=MIGfMA0GCSqGSIb3D...
        domain:
          type: string
          example: example.com
        selector:
          type: string
          example: dkimkey
        firstSeenAt:
          type: string
          format: date-time
          example: 2021-01-01T00:00:00Z
        lastSeenAt:
          type: string
          format: date-time
          example: 2021-01-01T00:00:00Z
    DomainSelectorPair:
      type: object
      required:
        - domain
        - selector
      properties:
        domain:
          type: string
          example: example.com
        selector:
          type: string
          example: s1